pyproject.toml
[tool.poetry]
name = "busride"
version = "0.0.1"
description = "Are we there yet?"
authors = ["Tony Gao <tonyzgao@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.54.1"
gino = "^1.0.0"
arq = "^0.19"
sqlalchemy-utils = "^0.36.5"
alembic = "^1.4.2"
psycopg2-binary = "^2.8.5"
uvicorn = "^0.11.5"
email_validator = "^1.1.0"
sentry-sdk = "^0.14.3"
gino-starlette = "^0.1.1"
Wikidata = "^0.7.0"
requests = "^2.27.1"
aioredis = "1.3.1"
pywikibot = "^6.6.3"
mwparserfromhell = "^0.6.3"
fastapi-login = "^1.7.3"
python-multipart = "^0.0.5"
feedgen = "^0.9.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
bandit = "^1.6.2"
flake8 = "^3.8.1"
rope = "^0.17.0"
watchgod = "^0.6"
taskipy = "^1.2.1"
pre-commit = "^2.4.0"
isort = {version = "^4.3.21", extras = ["pyproject"]}
pytest = "^6.2.5"
pytest-asyncio = "^0.17.2"

[tool.taskipy.tasks]
# Launch Commands for Local Development
app = "uvicorn app.main:app --reload"
worker = "arq app.worker.WorkerSettings --watch ./"
# Docker-Compose Tasks
compose-up = "docker-compose -f docker-compose.local.yml -f docker-compose.worker.yml -f docker-compose.yml up --build"
compose-down = "docker-compose -f docker-compose.local.yml -f docker-compose.worker.yml -f docker-compose.yml down --remove-orphans"

[tool.black]
line-length = 79
target-version = ['py38']
exclude = '''
/(
    \.git
  | .vscode
  | build
  | dist
  | .scripts
  | .git-crypt
  | services/.scripts
  | services/.vscode
  | core/.scripts
  | core/.vscode
)/
'''

[tool.isort]
line_length = 79
multi_line_output = 3
not_skip = "__init__.py"
use_parentheses = true
include_trailing_comma = true
force_grid_wrap = 0
known_third_party = ["alembic", "arq", "fastapi", "gino", "pydantic", "sentry_sdk", "sqlalchemy", "sqlalchemy_utils", "starlette"]